rename pkg to internal
