+++ internal/domain/common.go
