+++ internal/domain/mock/notification.go
@@ -3,7 +3,7 @@ package mock
 import (
 	"context"
 
-	"exchange/pkg/domain"
+	"exchange/internal/domain"
 )
 
 type NotificationService struct{}
